adding ability to download books
