+++ AH_BookStore_project/author/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AuthorConfig(AppConfig):
+    name = 'author'
