+++ AH_BookStore_project/pages/templates/book/book_detail.html
@@ -45,4 +45,5 @@
 	{% else %}
 		<img class="book_detail_page_image" src="{% static 'logo.png'%}">
 	{% endif %}
+	<a href="share/" target="_blank"><button>Send it to me</button></a>
 {% endblock content %}
