+++ AH_BookStore_project/pages/templates/pages/contact_us.html
@@ -3,6 +3,14 @@
 	Contact us
 {% endblock title %}
 {% block content %}
+	<table>
+		<tr>
+			<th>Whatsapp : </th><td>09100024177</td>
+		</tr>
+		<tr>
+			<th>Email : </th><td>ahmadihamed231@gmail.com</td>
+		</tr>
+	</table>
 	<form  method="post" action="." novalidate>
 		{{form.as_p}}
 		{% csrf_token %}
