+++ AH_BookStore_project/publisher/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+from .models import Publisher
+# Register your models here.
+
+class PublisherAdmin(admin.ModelAdmin):
+	list_display=("name",)
+	prepopulated_fields={"slug":("name",)}
+
+
+admin.site.register(Publisher,PublisherAdmin)
