+++ AH_BookStore_project/publisher/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PublisherConfig(AppConfig):
+    name = 'publisher'
